fixed join_node() in server, so it writes drbdctrl.res files again
the return code of load_conf() had been changed from
an integer error code to a list of error entries, but in
join_node(), there was still a comparison of that
return code with zero. Fixed by adding code in join_node()
that interprets the error list returned by load_conf().