(Experimental) fixes for multivolume deployments
Support for DRBD clients broke multivolume deployments,
because it required the specification of volumes on a
per-node basis. Drbdmanage created a volume configuration
only for volumes that had already been deployed, but
drbdadm expects to see a volume on all hosts.

Therefore, drbdmanage now creates a configuration that
will contain all the already-deployed volumes of the local
node and only peer-volumes of those volumes (any
peer-volumes, whether deployed or not, that are not
yet deployed on the local node, will be ignored).
The configuration file always pretends that remote
nodes have backend storage (using "/dev/null" as the
device path) to avoid an error when deploying a DRBD
client before the first node with backend storage finishes
deploying all volumes.