added "startup" command
for intuitive use; actually, any command can be
used to trigger D-Bus activation of the drbdmanage
server, but someone who is unfamiliar with that
principle may be looking for a specific startup
command.
Also, the response to "ping" is just "pong" or a
detailed D-Bus error message, which is
somewhat unexpected as an answer to a startup
command. For this reason, the startup command
has been implemented with better hints as to how
the server startup is attempted, and where to look
for the problem cause if the startup fails.