do not set the update flag on control-volume-less nodes
Normally, if the list of drbdmanage nodes changes, the update flag
is set on all peer nodes, to prompt those nodes to update their
drbdctrl.res file and to adjust the connections of the control volume
resource. Nodes without a control volume can skip all those steps,
so there is no reason to set the update flag on those nodes in the
first place.