updated primary/secondary role logic for snapshot restore
DrbdManager's primary_deployment() function determines
whether an assignment that is being deployed should serve
as the SyncSource for the same DRBD resource on other
nodes.
This function has updated logic now, so it should be able to
handle resources that have data based on a restore operation
of a snapshot that is not available on all nodes the resource
is assigned to.