DrbdManager class: experimental snapshot restore implementation
- in the DrbdManager class (drbdcore module):
  _deploy_volume() was restructured and split into multiple
  subroutines:
  _deploy_volume_actions() decides whether to create new
  (empty) DRBD volumes, or to restore the volumes from
  a snapshot by creating another snapshot of that snapshot
  and using that as the backend storage for the volume
  _deploy_volume_blockdev() creates a new block device
  for a new empty DRBD volume
  _deploy_volume_from_snapshot() creates a new block
  device by creating a snapshot of an existing snapshot;
  this is used for restoring snapshots
  _deploy_volume_metadata() creates DRBD metadata on
  a block device

Known problems:
Currently, if a new resource is based on a snapshot restore,
the node number for that resource on each node may be
different than the one that had been used on the resource
the snapshot was taken from. This must be changed, because
otherwise DRBD will refuse to use the existing metadata on
the (restored) snapshot volume.