utils: Added ssh_exec function
Currently, this is only used by the join command, but we will probably
make use of it in the future (e.g., satellite nodes).
While at it, clarified that the command is executed on a remote host and
that output shown comes from remote, and input is executed on remote.
This was requested on the drbd mailing list.