drbdmanage: Support for drbdsetup options
This adds support for drbdsetup options, namely "disk-options",
"net-options", and "resource-options" as follows:

There is new class "DrbdSetupOpts" which executes "drbdsetup" and
parses all valid options. This then generates drbdmanage subcommands
including description, type and range checks. For example, there is then
a subcommand "net-options", that can be called with "--max-buffers". As
for example net-options are allowed in the "common" section as well as
per resource, these commands then provide the according switches.

Setting max-buffers for a resource "abc" looks like this:

$ drbdmanage net-options --max-buffers 2048 --resource abc

Setting this option in the common section looks like this:
$ drbdmanage net-options --max-buffers 2048 --common

Additionally, there is always an "--unset-" option for every option that
can be specified. So, unsetting max-buffers for a resource "abc" looks
like this:

$ drbdmanage net-options --unset-max-buffers --resource abc

Options are stored in their own namespace in the according
propscontainer.