propscontainer: Added optional namespaces
Namespaces are separated  by '/', like Unix file paths, and can be used
to store and retrieve key/value pairs. E.g.: a key "foo", with value
"bar" within the namespace "test" is stored as:
/test/foo = bar

This allows to get only the properties that are currently of interest,
like all disk-options of a resource, without tricky filtering at the
call site.