Provide modify-config command and API
This adds a modify-config command with the following benefits:

- Local config files are avoided as much as possible. modify-config
writes its configuration to the control volume. As a consequence the
configuration is now distributed in the whole cluster. In rare
situations it might be necessary to provide a local config file
/etc/drbdmanaged.cfg in case the control volume itself is placed on a
non-standard VG. This configuration file has to provide a [LOCAL]
section. Allowed values include 'drbdctrl-vg' and 'extend-path'. If
possilbe a local configuration file should be avoided.

- modify-config: This starts an editor and allows to configure all the nodes
it knows about. Options can be set in the [GLOBAL] section and apply to
all nodes, or in [Node:xyz] sections. Values in node sections overwrite
global settings. Values specified in a local configuration file
overwrite all other settings.

- modify-config --node xyz: This allows to set more specific settings
including plugin settings (e.g., settings relevant for the ThinLV plugin in a
[Plugin:ThinLV] section).

- The plugin system got rewritten and now contains a plugin manager that
is able to instantiate plugins and query its settings. The plugin
manager is also responsible for plugin reconfiguration. If the
configuration did not change, reconfiguration is avoided. Otherwise, the
plugin's set_config method is called. It is then up to the plugin to
implement the required (re)configuration steps.

- A Dbus-API for setting the cluster/node configuration is provided.

- Configuration syntax as defined in:
https://docs.python.org/2/library/configparser.html

- All existing configuration files (drbdmanaged.conf + plugin specific
.conf files) are now ignored and can/should be removed. The reason for
the changed file suffix (.cfg instead of .conf) is to avoid confusing
the parser and to allow vim syntax highlighting (default vim knows how to
highlight .cfg files, but not .conf files)