Cluster-wide configuration and serial number generation changes
The cluster-wide configuration is now stored in a PropsContainer.
Since PropsContainers maintain their own serial number, but the
serial number is stored in the cluster-wide configuration,
which is not a PropsContainer itself, serial number generation has
been modified so that the root PropsContainer, currently the one
which holds the cluster configuration, creates a serial number
generator which it will use directly for updating serial numbers.
Whenever other objects request a new serial number from the
server, the server uses the root PropsContainer to create
one serial number per change, and also uses the serial number
generator to close the current change generation.
Both, the cluster configuration and the serial number generator,
can be found in the server's objects directory (_objects_root)