Added initial site concept
It is now allowed to set drbdmanage settings (e.g., deployer-plugin) per
site.

Sites are defined within [Site:abc] sections and are allowed to contain
the usual drbdmanage settings. [Node:def] sections are now allowed to
specify the option "site = abc". The hierarchy for settings is now
(starting with the lowest priority):
- [Global]
- [Site:]
- [Node:]
- drbdmanaged.cfg

This commit is also a preparation step to further integrate drbdsetup
options (net-options, disk-options, resource-options) in the new site
concept. For example it should be possible to set net-options that are
valid for a connection between siteA and siteB.