Read the current replication status of the control volume upon startup
If the control volume is already up and connected at the time the
drbdmanage server is started, then drbdmanage's quorum manager
must figure out which nodes are already connected to and replicating
the control volume. This can be achieved by not only tracking change
events, but also the initial connection/replication status of the
control volume.