provide saner VG defaults for plugins
VG names are split between the drbdctrl VG and the VGs plugins use,
which is basically a good idea. If the user sets the drbdctrl VG (e.g.,
via 'drbdctrl-vg' in drbdmanaged.cfg), so far the user also had to adapt
the plugin settings. It can be assumed, that plugins should use the same
(non standard) VG as the drbdctrl volume as long as the VG setting for a
plugin is not overwritten by the user.