sites: allow net-options per site, rewrite config generation
This commit allows network settings per site. For example if there are 4
nodes in your cluster (Node{A,B,C,D}) and NodeA and NodeB are part of
site A, and NodeC and NodeD are part of site B, then you can set
protocol C within sites A and B, and protocol A between sites A and B as
follows:
$ drbdmanage net-options --protocol C --sites 'A:A'
$ drbdmanage net-options --protocol C --sites 'B:B'
$ drbdmanage net-options --protocol A --sites 'A:B'

'--sites' follows a 'from:to' syntax, where currently 'from' and 'to'
have a symetric semantic. Setting an option from 'A:B' also sets this
option from 'B:A'.

The DRBD configuration generation code got extended to reflect the sites
concept:
- Depending on the configuration, there might now be multiple
connection-mesh sections.
- The generated configuration is minimal in the sense that
connections/connection-meshes are only generated for nodes that share a
connection. For example in NodeA's configuration, connections between
NodeB and NodeC do not show up.