tables: added support for utf-8
This is a patch I had in my queue for a too long time. The output on
utf-8 terminals looks really nice, but as long as we cannot disable it
(and fall back to ascii), we cannot enable it.

Too dangerous for old unix-beards running RHEL5/6 or something ancient.

So, the default is still 'ascii'.