Generate a new serial number upon removal of objects
The new dual control volume persistence layer commonly attempts
to open the volume with the newer serial number for reading and
the volume with the lower serial number for writing.
If both volumes have the same serial number, it always opens
the same one for reading and the other one for writing, thereby
essentially undoing any operation that does not update the serial
number, such as object removals. Updating the serial number
fixes the problem.