query-conf: Fixed querying remote nodes
Rewriting the logic in conffile.py introduced a regression:
The rewrite tries to write a minimal configuration (i.e., only the
configuration necessary for the node that writes the configuration).

This part of the code is also used for querying the configuration of
other nodes in the cluster. The result was that the configuration
generated was not complete.