debian: cleaner systemd integration
Now we use systemd integration if we find the according deb-helper. We
do _not_ add dh-systemd as build dependency because we still have to
support older LTS ubuntu releases that do not have dh-systemd.

If you want that integration, make sure dh-systemd is installed on your
build machines. That's what we do for drbd-utils.