client: allow stderr output in interactive mode
This fixes a bug where output to stderr was not written in interactive
mode. This whole stderr redirection dance is just needed to print error
messages more nicely in interactive mode. Without redirection the parser
itself prints its usage, which does not make sense, because we then
print the usage for the failed command.