fix error msgs on argparse failures
Currently the client did not print error messages correctly if argparse
failed because "stderr" was redirected. This whole crazy "stderr"
redirection was invented because of the "interactive" mode when a
command was executed with its '-h' flag. In that case the synopsis was
printed twice.

Avoid redirection at all and check for this specific situation instead.