fix help msgs in case of SyntaxExceptions
Most syntax exceptions are caught by argparse. We use SyntaxException
for (rare) situations where we have to check arguments closer. If this
fails, a function is called that then prints the correct synopsis. This
function expects a key "command". Until now, this key was not set and a
stack trace was the result.