Fix run_external_plugin() dbus command.
First, give meaningful error messages, instead of a generic
"not found" which is wrong most of the time.
Second, fix plain lying as in "Completed successfully".
Thirdly, in case of a problem, report some details, so that
fixing the problem becomes possible.
Fourthly, allow the plugin to actually return some data.