refactoring of storage plugin layer
This avoids lots of unnecessary code duplication and improves
maintainability in the long run.

Storage plugins are now derived from to classes:

StoragePlugin() - This class provides the interface and raises
NotImplemented errors if a plugin does not implement certain functions
(e.g., snapshots). While not strictly necessary, this class provides an
overview about the methods we expect developers to implement and some
documentation.

StoragePluginCommon() - This is the main parent class and provides the
default high level (and lower level) methods a plugin should implement.
Basically a harness/boilerplate that provides error handling and calls
the plugin specific methods

Usually, plugins only have to implement the lower level methods (the
ones starting with '_'), but are free to also overwrite the higher level
functions if necessary (e.g., lvm_thinpool).