build: add distclean target
"make release" now also asks if "distclean" was executed. I don't want
"distclean" to be a dependency of "release". While it is bad practice to
execute "make release" in a messy development directory, I don't want
deletion of files currently not added to the git index if a developer
executes "make release" in her/his development directory.