plugins: don't call set_plugin_conf for ext. plugins
Currently we call set_plugin_conf for every _loaded_ plugin. With the
introduction of external plugins this behavior has to change. External
plugins are currently one-shot plugins that are called, do their work,
and are done till called next time. They don't have a long living
configuration as for example the LVM storage plugin. Still, they are in
the list of loaded plugins to speed things up.

In this case it caused a bug where set_plugin_conf was called with a non
existing key, resulting in KeyError while accessing the persistent
storage.

This bug (only) affected projects making use of external plugins like
OpenStack and the soon to be released Docker driver.