zfs: fix regression introduced in 860426f
In 860426f a "traits" dictionary got introduced. In order to make use of
it, classes have to call init of their parent where this dictionary is
defined. This was not the case for the Zvol plugin, but it tried to set
values in that dictionary. The result is that loading the Zvol plugin
fails. I squashed that half finished change into Robert's commit, it is
me who is to blame.