Mark assignments as deployed before undeploying them
This should fix a case where a resource is assigned to a node,
and deploying the assignment was attempted, but then failed.
In this case, the resource configuration file is still present on
the target node, and a drbd resource or a block device may be
present, but the assignment is not marked as being deployed.
If an assignment is undeployed in such a situation, then it would
appear as having been undeployed already
(because in most cases, current state == target state == 0)
and the assignment may then be cleaned up immediately
without running any DrbdManager actions to undeploy the
assignment on the target node.
To resolve this issue, the assignment is marked as deployed now
before undeploying it, so the target node will always run the
DrbdManager actions to undeploy the assignment.