Add reasonable thin provisioning capability
To enable reasonable use of thinly provisioned storage,
the initial sync must be avoided, as well as full syncs
when adding another node that uses thinly provisioned
storage too.

Mixing thin volumes on one node and fat volumes on another
(for the same resource/volume) is supported, but may
cause the thin volumes to allocate their entire capacity
in certain cases.

For set-gi, drbdmeta is used directly instead of going through
drbdadm

drbdadm requires a resource with at least one connection
for running the set-gi command because of some additional
operations it could carry out.
That is a problem whenever a resource is only assigned to
a single node, and this is difficult to change in drbdadm,
therefore use drbdmeta instead of drbdadm.