add-node: start satellite
So far "add-node -l" only printed that drbdmanaged should be started on
the satellite node. Now we send it an activation too. The command sent
via ssh is "drbdmanage restart -q". That is also what is printed on
"howto-join".

Here we have to differentiate two situations:

- Not systemd socket activated: Everything is fine, "restart" brings up
drbdmanaged and it binds the satellite TCP port.

- Systemd socket activated: We execute a drbdmanage command, namely
"restart", which triggeres DBUS activation. Drbdmanaged has no clue that
it should be socket activated and tries to start a TCP server that does
not make use of systemd socket activation. This fails as the port is
already in use by sytemd. That is fine, just let it fail. The control
node eventually sends a command to the satellite which then gets started
by systemd socket activation.