utils.py: new get_free_number() implementation
The new implementation handles various corner-cases correctly
that the previous implementation did not support.
With the previous implementation, it was required that all numbers
in the list of occupied numbers were in the range from
min_nr to max_nr. The new implementation avoids this limitation
and should therefore help to prevent future implementation errors
when trying to allocate free numbers.