server: add DBUS tracer/replay
So far it was a bit cumbersome to get reproducible test cases out of
complex setups/stacks (e.g., OpenStack). Especially, without knowledge
about the project itself (OS, OpenNebula, OpenAttic, ...). Recording the
DBUS session and replaying should help drbdmanage developers.

This is an attempt to solve this on a general level, as much driver
agnostic as possible.

The first step is to enable tracing:
drbdmanage dbus-trace --start
Then you produce dbus calls and finally top tracing with
drbdmanage dbus-trace --stop

If successful, this returns a path on the server where you find a python
script that can be executed for replay.

Having all the commands executed in a reproducible way is nice, but only
half of the medal. Actually, timing is as important. "dbus-trace" also
records timing information. While this is never accurate, because
resource setup times may vary, it is better than nothing. Sleeping is
enabled by setting by setting the environment variable "DMDT_SLEEP".

It also allows single stepping which is enabled by setting the
environment variable "DMDT_STEP".