Introduce single leader
So far we had:
- control nodes: access to control volume, every control node was
responsible for some satellite nodes.

- satellite nodes: got a copy of the control volume via TCP/IP from
their respective control node.

This has several architectural and performance issues:
- We had to store which control node was responsible for which satellite
node.

- Handover of an satellite to a different control node was a multi-step
process. Closing the TCP connection, storing to the control volume that
now another control node is responsible for the satellite, the new
control node reads back that information and initiates a TCP connection.

- Having multiple control nodes that race for access to the control
volume is slow.

- Satellites were second class citizens. All operations that would have
altered the state of the control volume were not allowed, like "create
resource" -> "Operation not allowed on satellite node".

The new design is as follows:
One "leader" that has access to the control volume and is in DRBD
Primary role. All other nodes are in the role of a satellite and the
leader and the satellites communicate via TCP.

The control volume can still be replicated to several nodes in the
cluster. Every node that has access to the control volume is a
"potential leader". If the old leader vanishes, one of the potential
leaders becomes the current leader.

- Leader: A node with access to the replicated control volume, currently
DRBD Primary.
- Potential Leader: A node with access to the replicated control volume,
currently DRBD Secondary. In this role a nodes otherwise acts like a
satellite (ie. gets a copy of the control volume via TCP).
- Satellite: A node without access to the control volume. Always
satellite.

Leader election:
On startup nodes check if they have access to the control volume. If
not, they become satellites. Nodes that have access to the control
volume "race" to become DRBD Primary and therefore the current leader.
Only one wins (ensured by DRBD). If a leader vanishes and there is no
current leader, potential leaders start a reelection.

(Re)election strategy:
- 1 node cluster: obvious
- 2 node clusters: Wait until the control volumes are DRBD Connected.
Only if that is the case, start race for Primary. That is necessary to
avoid split brains.
- >3 node clusters: Quorum

In a two node cluster this could have unexpected consequences. One node
crashes and is permanently lost. The other node would not even be able
to remove that node from the cluster as it would wait forever to connect
to the second node. In such a case someone should use "drbdmanage
reelect --force-win", check if the node became Primary, delete the second
node, "drbdmanage restart -q".

Nodes in the role of a satellite are now allowed to execute every
command. This makes sense as now N-1 nodes are in the role of a
satellite (the pure satellites and the potential leaders). Such nodes
forward the request to the current leader.

To check the current role of a node, use "drbdmanage role".

Limitations/Changes:
Nodes are now very strict about the point in time when they are ready.
They have to run the leader election and if they lose, they have to wait
for an initial contact from the leader (otherwise they would not be able
to forward requests). This takes time, ~5 seconds. If a node is not yet
ready, it responds with a special error code that tells the client to
retry later. The commandline client handles that transparently. Other
projects like OpenStack will need updates. Currently, even the
ClientHelper class lacks that update. This will be fixed until the next
official release. While startup got slower, the rest feels a lot more
responsive.

We still send copies of the control volume (json + bz2). This could be
improved, but is currently no real limitation and optimizing that is not
worth it (see Next Steps).

Next Steps:
Stop sending the control volume and then let nodes react to it. Move the
decision logic to the leader. The rest of the nodes will become "light
satellites". The get simple commands like "create backend device foo"
and execute that simple step (and report back). It is up the leader,
and only to the leader, how to react (success, retry, rollback,...).