make use of the new features in "drbdadm adjust"
"drbdadm" recently got some nice features that reduce some logic
required in drbdmanage itself. We can remove connect/attach and call
adjust directly. This helps in scenarios where an admin manually fiddled
around on the command line and it should make the whole process of
resource creation/adjustment more stable.

This changes the internal logic, but keeps the old interface.

This requires functions that are available in drbd-utils>=8.9.7.