systemd: disable socket activation
So far we had a strict separation of control nodes (no TCP server) and
satellite nodes (TCP server). There it made sense to provide a socket
unit hat was used on the satellites.

With the new architecture things are more dynamic. The leader nodes also
has a running TCP server and nodes dynamically switch their role. The
service unit fits all, there is no need for a separate socket unit.

This commit disables the socket unit, as it made troubles and just
confuses users which unit file to enable in which situation.

This removes the socket unit for package building/install, but keeps the
file itself and the code in proxy.py. This might be completely removed
in the future.