single leader: pseudo fence if isolated
So far we could have the following scenario (e.g., 3 node cluster):
- A: leader, B and C: potential leader, currently satellites.
- A gets separated from the network for whatever reasons. It keeps being
  DRBD Primary.
- B and C detect that and elect a new leader.
- A comes back: Split brain.

Now A detects that and tries to become DRBD Secondary and to down the
control volume.

If that succeeds, the old leader (A in our example) enters reelection
and brings up the control volume with "--discard-my-data" as there might
have been already writes on the new leader.

If Secondary && down do not succeed, the old leader "fences" itself
(exit 1).