Make update-pool a cluster wide operation
A long time ago update pool was a simple operation where nodes checked
their local pool status and wrote it to the control volume.

This assumption broke when we introduced "satellites" and becomes
visible with the single leader concept where N-1 nodes are satellites.

This makes "update-pool" a cluster wide operation. Satellites forward
the command to their leader, and the leader starts a round of updates.

Whenever the ping service detects that a node is now pingable, it
collects the pool status from that node once.