server: remove invalid drbdsetup opts on startup
DRBD Manage parses valid options from "drbdsetup" at runtime and
generates the according set and unset commandline options. If you set an
options that was valid in an older version of drbdsetup (e.g.
unplug-watermark), that was written to the control volume. Later, when
you upgrade your "drbdsetup", the according unset-option is never
generated and it was impossible to get rid of a deprecated option.

Now we check on startup if there are any options set that are not valid
anymore and if so, we drop them.