ping service: Don't hang if nodes leave
Recap: The "ping service" is essential, as the leader sends pings to the
satellites so that they know who their leader is. For example satellites
don't consider themselves successfully started if they did not have
contact from their leader.

So far we tried to send messages to satellites unconditionally. Pings as
well as work load. This had the problem that if a node leaves for
whatever reason, we run into TCP timeouts and the cluster feels
unresponsive because the pings are scheduled too fast. Same for
nv/nr/...

This commit does the following:
- Make the ping period higher than the TCP timeout (and shorten the TCP
  timeout).
- If a node is not ping-able, set its state to "False".
- Send pings/work load only to nodes that are "True".
- Nodes that land on the "False list", get integrated by a reintegration
  service (every 30s). If there are multiple unreachable nodes on the
  "False list", one node is randomly picked every 30s.
- Whenever a leader is elected, there is a grace period of 2 minutes.
  Within this period, the described mechanism is disabled. This helps on
  startup where we don't want to set nodes on the "False list" too soon.
- If an admin brings back a node, she does not has to wait for 30s, a
  "resume-all" cleans the "False list".