satellites: fix startup if node is pure satellite
So far startup on satellites that never had access to the control volume
(locally) was only possible by coincidence when the leader sent a
message containing the control volume to that satellite. The satellite
got a ping from the leader, but then tried to look up the leader's IP in
the control volume it never received in the first place to initially
request the current control volume from the leader.

That broke when we changed the semantic from "leader sends initial
control volume" to "satellites request the control volume whenever they
need it". From the leader pings we know the leaders IP and we can then
use that IP to request the control volume without a lookup in the non
existing nodes list.