don't recycle minor numbers early
So far we recycled minor numbers as soon as possible. If a volume got
deleted, the next volume recycled the minor number of the old one. There
might be situations where drbdmanage thinks the volume is gone, but in
fact it isn't. Increase minor numbers until we hit the maximum value,
and then try to recycle ones that got free.