auto deploy: add limitation to sites
This implements a new dbus API call (auto_deploy_site) and corresponding
client options (--site, -s) for "new-volume", and "deploy". If a site is
set, only nodes from that site are considered for deployment. This makes
it possible to have a cluster of let's say 4 nodes, where storage is
strictly separated between a "customer" site and a "internal" site. This
of course needs a site configuration (drbdmange edit-config).