single leader: avoid Primary after fence
So far it was possible to switch the second node in a 2 node cluster to
Primary under this conditions:
- There was data divergence on the ctrl volume
- The node tried to connect, but went to StandAlone
Usually we use "wait-connect-resource" to check that the ctrlvols are
indeed connecting befor we start the race to become Primary (for leader
election). This works as expected if the node is "Connecting", but fails
if the node is "Standalone". In the first case wait-connect-resource
returns a non zero exit status, but in the second case returns with a
zero exit status.

This can be fixed by appending " --wait-after-sb=yes". That old behavior
isn't effected, but it also fails (and then waits) in the StandAlone
case.