enable quorum tracking on all nodes with ctrlvol
So far we disabled quorum tracking when the node was a satellite, no
matter if it had a ctrlvol or not. This had a usability problem where
nodes currently offline were shown as "ok" on all nodes not the current
leader (because they were satellites). This is not necessary and
confuses users. Now we track the node state on all nodes that have
access to the ctrlvol, no matter if leader or currently a satellite with
access to ctrlvol. Actually they just need access to "drbdsetup events2"
tracking.