avoid creation of broken res files
In hard to reproduce situations we ended up with .res files that did not
have connection/on sections.

This happened because the nodes were not in the "interesting" list,
which was caused by a race on the objects_root data structure. The class
that finally creates the res files got initialized with a set of server
and client nodes. Then, during creation, and in some paths, the
algorithm had to check if the instance node is part of given subsets of
the overall nodes. In that case it queried the objects_root for what the
instance node object is, and then did not find the new self (current
point in time) in the old list (class object creation time).

Here an example where the new "foo" did not find is old object:
 _gen_mesh_conf, inst_node <drbdmanage.drbd.drbdcore.DrbdNode object at 0x7ff4dde7c210>, foo
 _gen_mesh_conf, mesh_node:<drbdmanage.drbd.drbdcore.DrbdNode object at 0x7ff4dc14aa50>, bar
 _gen_mesh_conf, mesh_node:<drbdmanage.drbd.drbdcore.DrbdNode object at 0x7ff4dc055fd0>, foo