wait-for/openstack: fix calls on satellite
So far we had a semantic mismatch between the cli and the wait-for
plugins. If you called for example "list-assignments" on the
command line, the client fetched the current state from the leader node.
Background: Usually the leader only forwards the new control volume to
the satellites if it is necessary (e.g., they get a new assignment). This
had the disadvantage that only the leader saw the complete state,
satellites could have had an outdated view. That's why ctrlvol fetching
was introduced for "list commands" on the satellites.

The wait-for plugin did not fetch the ctrlvol, but just looked into the
(maybe outdated) internal structures of the drbdmanage server and
therefore did not see all changes.

This patch fetches the current state in the plugin and allows e.g.,
cinder to be executed on a satellite node. So far this only worked if
the cinder node was the drbdmanage leader node.