client: fix bug in inline comments for ConfigParser
In order to help the user, we have some key/value pairs that look like
that:
"# key = value # comment"

If the user removes the first '#', then the value also contains the
comment (e.g., "value # comment"). This is documented [1]. Unfixed for
"old" Python, and fixed somewhere in the middle of Python 3...

We have such an inline comment in one place. Fix that by introducing a
newline and a "# ^^ comment" line. Stripping these comments would have
touched multiple call-sites and would have required regenerating tuple
lists.

[1] https://bugs.python.org/issue14590