server: reset grace period on create_node
We constantly ping the satellites (so that they know who their leader is
in case of a leader switch). As there might be TCP timeouts, satellites
that fail to answer are set to a "blacklist". Then they do not get
contacted that frequently and are integrated later.

For new nodes this might be problematic if the do not answer immediately
because they are still in their startup. So reset the grace period on
every create_node which should give new nodes enough time.