avoid shutting down satellites immediately
When a satellite gets removed, the node entry is removed from the
control volume and we set the node on a list of nodes for proposed
shutdown. If the satellite does not have anything to contribute anymore,
we move it to a second list, which gets processed periodically and
finally we send nodes on that second list a shutdown signal via TCP.

We moved nodes to the second list only if they did not have any
contributions but not if they did not answer at all. So assume a
satellite that got added but never answered anything (e.g., did not
successfully start up). On remove-node it was put on the first list, but
as it never answered, it was never put on the second list. A new
"add-node" for the same node(name) that was successful and where the
satellite answered, *then* put it on the second list and killed it
immediately.

Fix this by moving nodes that are on the proposed list to the final list
if they do not have anything to say anymore OR if the do not answer at
all.