make cluster initialization more reliable
- Disable quorum on add-node: Having the quorum enabled during init made
troubles on adding the 3rd node. Disable the quorum temporarily. It is
(re)enabled on connect for these nodes later anyways.

- Remove any existing .drbdctrl/drbdctrl.res resources/files on join. A
"join" is executed on a node that joins a new cluster by definition, so
it is save to remove the old configuration.

- Start the server and wait for its startup on "join". Otherwise we run
into dbus-timeouts frequently.

- Send a final "poke" and "update-pool" so that the cluster has current
information. This only works via autojoin/ssh. So this is the supported
way, therefore make sure the node executing "add-node" has root ssh
access to the joining node.

- Do not reelect on join. This does not make sense as we startup the
server anyways and a fresh node that should join will be in the role of
a satellite anyways.