fix too short timeout for leader/satellite communication
The leader node sends a lot of TCP messages to the satellites. For
example it sends pings in a "ping service" frequently to inform (new)
satellites who their leader is and continues to do that to check if a
satellite is still there.

If there is no existing socket for a satellite, the leader creates it.
In order to avoid long timeouts, we did:
socket.create_connection((peer_ip, port), timeout=2)

It was assumed that this timeout is only for the connection initiation,
but in fact it sets the general socket timeout.

So satellites had essentially only 2 seconds time to finish their work
(eg., creating the resource file, creating meta data, adjusting the
 resource). If that took longer, which can happen in busy clusters, we
had the following situation:

Leader to satellite communication was always successful (e.g. "create
resource foo"). The satellite did its work (i.e., created the drbd
resource), but was not fast enough to answer the new state back. The
leader's receive call timed out, and therefore the new and successful
state was never accounted. In the next send to the satellite, we sent
the old state ("create resource") again, the satellite tried then to
remove things and recreate things. Even if successful, the satellite
would not be able to communicate back because of the too short timeout.

Users saw a happy drbd, and drbdmanage complaining about pending
actions.

Fix that by:
- using a short timeout in general. This keeps for example the ping
service happy without blocking the leader.
- use a sufficiently long timeout for commands that are expected to take
longer.