fix locking between satellites/leader
This is also bug introduced when switching to implicitly switching to a
multithreaded design due to the multithreaded TCP server.

So far this could happened:
The leader writes a pending action it then wants to send to a satellite.
Concurrently the satellite could have requested a new, actually old,
state of the control volume due to a "list" command which would
overwrite its local data. This would then be sent back to the leader,
even if the local action was successfully.

L: pending action for satellite
L: send ctrl-vol with pending action to satellite
S: execute the action (e.g., create resource)
S: list command which retrieves the old control volume from leader
S: send back the old control volume with the pending actions
L: receive the old control volume
L: retry action

This could be triggered on satellites in busy clusters that also do list
operations. We say that in combination with OpenNebula. In that case the
driver already executed "list-commands" to see if the deployment was
successful. It can be manually triggered by executing a list-command in
a while-true loop while creating resources on the leader.

The leader now does not allow list-actions while it is in a
transaction.