storage backends add notes about remove_snapshot()
The idea was that every plugin can provide/overwrite a method hat is
called to remove a snapshot. In the current design all of them called
remove_blockdevice(). The storage core actually already called
remove_blockdevice(). Strictly speaking this is a bug, because it
bypasses the plugins remove_snapshot(). As all plugins only provide a
wrapper, just add a note and don't touch all the existing plugins. This
can be fixed when we need it, but document it for now.